uniformize traits
