+++ benchmarks/benches/attention.rs
@@ -5,7 +5,7 @@ fn main() {
     for problem in attention::problems() {
         for strategy in attention::strategies() {
             println!("---- {} / {} ----", strategy.label, problem.label);
-            match attention::run(strategy.id, problem.id, 10) {
+            match attention::run(&strategy.id, &problem.id, 10) {
                 Ok(samples) => {
                     let durations = BenchmarkDurations {
                         timing_method: TimingMethod::System,
