+++ benchmarks/benches/conv2d.rs
@@ -1,199 +1,20 @@
-use cubecl::{
-    Runtime,
-    benchmark::{Benchmark, ProfileDuration, TimingMethod},
-    client::ComputeClient,
-    future,
-    prelude::*,
-    std::tensor::TensorHandle,
-};
-use cubek::{
-    convolution::{
-        self as convolution, AcceleratedTileKind, ConvAlgorithm, ConvolutionArgs,
-        ConvolutionInputs, Strategy,
-    },
-    matmul::definition::{MatmulElems, MatmulPrecision, MatrixPrecision},
-    random::random_uniform,
-    std::InputBinding,
-};
-use std::marker::PhantomData;
-
-type LhsG<MP> = <<MP as MatmulPrecision>::Lhs as MatrixPrecision>::Global;
-type LhsS<MP> = <<MP as MatmulPrecision>::Lhs as MatrixPrecision>::Stage;
-type RhsG<MP> = <<MP as MatmulPrecision>::Rhs as MatrixPrecision>::Global;
-type AccG<MP> = <<MP as MatmulPrecision>::Acc as MatrixPrecision>::Global;
-type AccR<MP> = <<MP as MatmulPrecision>::Acc as MatrixPrecision>::Register;
-
-impl<R: Runtime, MP: MatmulPrecision> Benchmark for Conv2dBench<R, MP> {
-    type Input = (TensorHandle<R>, TensorHandle<R>, TensorHandle<R>);
-    type Output = ();
-
-    fn prepare(&self) -> Self::Input {
-        let client = R::client(&self.device);
-
-        let input = TensorHandle::empty(
-            &client,
-            self.input_shape.to_vec(),
-            LhsG::<MP>::as_type_native_unchecked(),
-        );
-        random_uniform(
-            &client,
-            0.0,
-            1.0,
-            input.clone().binding(),
-            LhsG::<MP>::as_type_native_unchecked().storage_type(),
-        )
-        .unwrap();
-        let weight = TensorHandle::empty(
-            &client,
-            self.weight_shape.to_vec(),
-            RhsG::<MP>::as_type_native_unchecked(),
-        );
-        random_uniform(
-            &client,
-            0.0,
-            1.0,
-            weight.clone().binding(),
-            RhsG::<MP>::as_type_native_unchecked().storage_type(),
-        )
-        .unwrap();
-        let bias = TensorHandle::empty(
-            &client,
-            vec![self.bias_shape],
-            AccG::<MP>::as_type_native_unchecked(),
-        );
-        random_uniform(
-            &client,
-            0.0,
-            1.0,
-            bias.clone().binding(),
-            AccG::<MP>::as_type_native_unchecked().storage_type(),
-        )
-        .unwrap();
-
-        (input, weight, bias)
-    }
-
-    fn execute(&self, (input, weight, bias): Self::Input) -> Result<(), String> {
-        let client = R::client(&self.device);
-        let [n, _, h_in, w_in] = self.input_shape;
-        let [c_out, _, k_h, k_w] = self.weight_shape;
-        let [s_h, s_w] = self.args.stride;
-        let [p_h, p_w] = self.args.padding;
-        let [d_h, d_w] = self.args.dilation;
-
-        let h_out = (h_in + 2 * p_h - d_h * (k_h - 1) - 1) / s_h + 1;
-        let w_out = (w_in + 2 * p_w - d_w * (k_w - 1) - 1) / s_w + 1;
-
-        let elems = MatmulElems::new_deprecated::<MP>();
-
-        let out: TensorHandle<R> =
-            TensorHandle::empty(&client, vec![n, c_out, h_out, w_out], elems.acc_global);
-
-        convolution::launch_ref::<R, 2>(
-            &Strategy::Inferred {
-                algorithm: ConvAlgorithm::SimpleSyncCyclic,
-                tile_kind: AcceleratedTileKind::Cmma,
-            },
-            &self.client,
-            ConvolutionInputs::Forward {
-                input: InputBinding::Normal(input.binding(), elems.lhs_global),
-                weight: InputBinding::Normal(weight.binding(), elems.rhs_global),
-                bias: Some(InputBinding::Normal(bias.binding(), elems.acc_global)),
-                out: out.binding(),
-            },
-            self.args.clone(),
-            elems,
-        )
-        .map_err(|it| format!("{it:?}"))?;
-        Ok(())
-    }
-
-    fn name(&self) -> String {
-        let client = R::client(&self.device);
-        format!(
-            "{}-conv2d-{}-{}-{}-{}",
-            R::name(&client),
-            LhsG::<MP>::as_type_native_unchecked(),
-            LhsS::<MP>::as_type_native_unchecked(),
-            AccR::<MP>::as_type_native_unchecked(),
-            AccG::<MP>::as_type_native_unchecked(),
-        )
-        .to_lowercase()
-    }
-
-    fn sync(&self) {
-        future::block_on(self.client.sync()).unwrap()
-    }
-
-    fn profile(&self, args: Self::Input) -> Result<ProfileDuration, String> {
-        self.client
-            .profile(|| self.execute(args), "conv-bench")
-            .map(|it| it.1)
-            .map_err(|it| format!("{it:?}"))
-    }
-}
-
-#[allow(dead_code)]
-pub struct Conv2dBench<R: Runtime, MP> {
-    input_shape: [usize; 4],
-    weight_shape: [usize; 4],
-    bias_shape: usize,
-    args: ConvolutionArgs<2>,
-    device: R::Device,
-    client: ComputeClient<R>,
-    _phantom: PhantomData<MP>,
-}
-
-#[allow(dead_code)]
-fn run<R: Runtime, MP: MatmulPrecision>(device: R::Device) {
-    let client = R::client(&device);
-    let batch_size = 16;
-
-    let bench1 = Conv2dBench::<R, MP> {
-        input_shape: [batch_size, 3, 227, 227],
-        weight_shape: [96, 3, 11, 11],
-        bias_shape: 96,
-        args: ConvolutionArgs {
-            stride: [4, 4],
-            padding: [0, 0],
-            dilation: [1, 1],
-        },
-        client: client.clone(),
-        device: device.clone(),
-        _phantom: PhantomData,
-    };
-
-    let bench2 = Conv2dBench::<R, MP> {
-        input_shape: [batch_size, 4, 256, 256],
-        weight_shape: [64, 4, 8, 8],
-        bias_shape: 64,
-        args: ConvolutionArgs {
-            stride: [1, 1],
-            padding: [0, 0],
-            dilation: [1, 1],
-        },
-        client: client.clone(),
-        device: device.clone(),
-        _phantom: PhantomData,
-    };
-
-    for bench in [bench1, bench2] {
-        println!(
-            "input: {:?} weight: {:?}",
-            bench.input_shape, bench.weight_shape
-        );
-        println!("{}", bench.name());
-        println!("{}", bench.run(TimingMethod::System).unwrap());
-    }
-}
-
-#[allow(unused)]
-fn run_benches<R: Runtime, MP: MatmulPrecision>() {
-    let client = R::client(&Default::default());
-
-    run::<R, MP>(Default::default());
-}
+use benchmarks::conv2d;
+use cubecl::benchmark::{BenchmarkDurations, TimingMethod};
 
 fn main() {
-    run_benches::<cubecl::TestRuntime, half::f16>();
+    for problem in conv2d::problems() {
+        for strategy in conv2d::strategies() {
+            println!("---- {} / {} ----", strategy.label, problem.label);
+            match conv2d::run(&strategy.id, &problem.id, 10) {
+                Ok(samples) => {
+                    let durations = BenchmarkDurations {
+                        timing_method: TimingMethod::System,
+                        durations: samples.durations,
+                    };
+                    println!("{durations}");
+                }
+                Err(err) => println!("error: {err}"),
+            }
+        }
+    }
 }
