+++ benchmarks/benches/gemm.rs
@@ -1,454 +1,20 @@
-use crate::matmul::launch::launch_ref;
-use cubecl::{
-    benchmark::{Benchmark, BenchmarkComputations, BenchmarkDurations, TimingMethod},
-    future,
-    prelude::*,
-    std::tensor::TensorHandle,
-    zspace::shape,
-};
-use cubek::{
-    matmul::{
-        self as matmul,
-        components::{stage::PartitionBuffering, tile::TileMatmulKind},
-        definition::{
-            LoadingPrecomputeStrategy, MatmulElems, MatmulPrecision, MatmulProblem,
-            TilingBlueprint, TilingScheme,
-        },
-        launch::Strategy,
-        routines::{
-            BlueprintStrategy, TileSizeSelection, double_buffering::DoubleBufferingArgs,
-            double_unit::DoubleUnitSelectionArgs, ordered_double_buffering::OrderedSelectionArgs,
-            simple::SimpleArgs, simple_unit::SimpleUnitSelectionArgs,
-        },
-    },
-    random::random_uniform,
-    std::{
-        InputBinding, MatrixLayout, StageSize,
-        cube_count::{CubeCountStrategy, GlobalOrder, HypercubeBlueprint},
-    },
-};
-use std::collections::BTreeMap;
-
-impl<R: Runtime> Benchmark for GemmBench<R> {
-    type Input = (TensorHandle<R>, TensorHandle<R>);
-    type Output = ();
-
-    fn prepare(&self) -> Self::Input {
-        let client = R::client(&self.device);
-
-        let mut lhs = TensorHandle::empty(
-            &client,
-            vec![self.b, self.m, self.k],
-            self.dtypes.lhs_global,
-        );
-        if self.tl {
-            let len = lhs.metadata.rank();
-            lhs.metadata.strides_mut().swap(len - 2, len - 1);
-        }
-        random_uniform(
-            &client,
-            0.0,
-            1.0,
-            lhs.clone().binding(),
-            self.dtypes.lhs_global,
-        )
-        .unwrap();
-
-        let mut rhs = TensorHandle::empty(
-            &client,
-            vec![self.b, self.k, self.n],
-            self.dtypes.rhs_global,
-        );
-
-        if self.tr {
-            let len = rhs.metadata.rank();
-            rhs.metadata.strides_mut().swap(len - 2, len - 1);
-        }
-
-        random_uniform(
-            &client,
-            0.0,
-            1.1,
-            rhs.clone().binding(),
-            self.dtypes.rhs_global,
-        )
-        .unwrap();
-
-        (lhs, rhs)
-    }
-
-    fn execute(&self, (lhs, rhs): Self::Input) -> Result<Self::Output, String> {
-        let client = R::client(&self.device);
-        let out = TensorHandle::empty(
-            &client,
-            vec![self.b, self.m, self.n],
-            self.dtypes.acc_global,
-        );
-
-        match launch_ref(
-            &self.strategy,
-            &self.client,
-            InputBinding::Normal(lhs.binding(), self.dtypes.lhs_global),
-            InputBinding::Normal(rhs.binding(), self.dtypes.lhs_global),
-            out.clone().binding(),
-            &mut self.dtypes.clone(),
-        ) {
-            Ok(_) => Ok(()),
-            Err(err) => Err(format!("{err:?}")),
-        }
-    }
-
-    fn name(&self) -> String {
-        let client = R::client(&self.device);
-
-        format!(
-            "{}-matmul-Lhs<{}-{}-{}>-Rhs<{}-{}-{}>-{}-{}-{}",
-            R::name(&client),
-            self.dtypes.lhs_global,
-            self.dtypes.lhs_stage,
-            self.dtypes.lhs_register,
-            self.dtypes.rhs_global,
-            self.dtypes.rhs_stage,
-            self.dtypes.rhs_register,
-            self.dtypes.acc_register,
-            self.dtypes.acc_global,
-            self.strategy
-        )
-        .to_lowercase()
-    }
-
-    fn sync(&self) {
-        future::block_on(self.client.sync()).unwrap()
-    }
-
-    fn profile(&self, args: Self::Input) -> Result<cubecl::benchmark::ProfileDuration, String> {
-        self.client
-            .profile(|| self.execute(args), "matmul-bench")
-            .map(|it| it.1)
-            .map_err(|err| format!("{err:?}"))
-    }
-}
-
-#[allow(dead_code)]
-struct GemmBench<R: Runtime> {
-    b: usize,
-    m: usize,
-    k: usize,
-    n: usize,
-    tl: bool,
-    tr: bool,
-    strategy: Strategy,
-    device: R::Device,
-    client: ComputeClient<R>,
-    dtypes: MatmulElems,
-}
-
-#[allow(unused)]
-fn entry(m: usize, n: usize, k: usize) -> (usize, usize, usize, usize) {
-    let expected = 2 * 6144 * 6144 * 6144;
-    let num_ops = 2 * m * n * k;
-
-    let b = Ord::max(expected / num_ops, 1);
-    let b = 2usize.pow(b.ilog2());
-    let b = Ord::min(4096, b);
-
-    (b, m, n, k)
-}
-
-#[allow(dead_code, clippy::single_element_loop)]
-fn run<R: Runtime, MP: MatmulPrecision>(device: R::Device, strategy: Strategy) {
-    for tl in [MatrixLayout::ColMajor, MatrixLayout::RowMajor] {
-        for tr in [MatrixLayout::ColMajor, MatrixLayout::RowMajor] {
-            for (b, m, n, k) in [
-                (2, 1, 4096, 4096), // entry(8192, 8192, 8192),
-                                    // entry(6144, 6144, 6144),
-                                    // entry(4096, 4096, 4096),
-                                    // entry(2048, 2048, 2048),
-                                    // (2, 1024, 1024, 1024),
-                                    // entry(512, 512, 512),
-                                    // entry(64, 1024, 64),
-                                    // entry(32, 1024, 32),
-                                    // entry(10, 1024, 10),
-                                    // entry(64, 64, 1024),
-                                    // entry(32, 32, 1024),
-                                    // entry(10, 10, 1024),
-                                    // entry(1024, 64, 64),
-                                    // entry(1024, 32, 32),
-                                    // entry(1024, 10, 10),
-                                    // (16, 1, 2048, 8192),
-                                    // (16, 1, 4096, 4096),
-                                    // (1, 512, 512, 512),
-                                    // (2, 8192, 8192, 1), // Outer
-                                    // (2, 8192, 1, 8192), // MatVec
-                                    //(2, 1, 8192, 8192), // VecMat
-            ] {
-                println!("-------------------");
-
-                let problem = MatmulProblem::from_parameters(
-                    m,
-                    n,
-                    k,
-                    shape![b],
-                    shape![b],
-                    tl,
-                    tr,
-                    MatrixLayout::RowMajor,
-                    None,
-                    None,
-                    MatmulElems::new_deprecated::<MP>().as_global_elems(),
-                    AddressType::U32,
-                );
-                let _ = run_one::<R, MP>(device.clone(), strategy.clone(), &problem);
-            }
-        }
-    }
-}
-
-#[allow(dead_code)]
-fn run_one<R: Runtime, MP: MatmulPrecision>(
-    device: R::Device,
-    strategy: Strategy,
-    problem: &MatmulProblem,
-) -> Result<(BenchmarkDurations, f64), String> {
-    let client = R::client(&device);
-    let b = problem.num_batches();
-    let m = problem.m;
-    let n = problem.n;
-    let k = problem.k;
-    let tl = matches!(problem.lhs_layout, MatrixLayout::ColMajor);
-    let tr = matches!(problem.rhs_layout, MatrixLayout::ColMajor);
-
-    let bench = GemmBench {
-        b,
-        m,
-        k,
-        n,
-        tl,
-        tr,
-        client: client.clone(),
-        device: device.clone(),
-        strategy: strategy.clone(),
-        dtypes: MatmulElems::new_deprecated::<MP>(),
-    };
-    println!("b: {b} m: {m} n: {n} k: {k}, tl {tl}, tr {tr}");
-    println!("{}", bench.name());
-
-    match bench.run(TimingMethod::System) {
-        Ok(val) => {
-            let flops = 2 * b * m * n * k;
-            let computed = BenchmarkComputations::new(&val);
-            let tflops = flops as f64 / (computed.median.as_secs_f64() * 1e12);
-            println!("TFLOPS: {tflops}");
-            println!("Times: {val}");
-            Ok((val, tflops))
-        }
-        Err(err) => {
-            println!("{err:?}");
-            Err(err)
-        }
-    }
-}
-
-#[allow(unused, clippy::single_element_loop)]
-// This function should be customized to help build a proper selector that reduces the number of
-// possibilities.
-fn run_grid_search<R: Runtime, MP: MatmulPrecision>() {
-    let client = R::client(&Default::default());
-
-    let mut algos = BTreeMap::<u64, (BenchmarkDurations, TilingBlueprint, f64)>::new();
-
-    let (b, m, n, k) = (4096, 10, 64, 10);
-    let tl = MatrixLayout::RowMajor;
-    let tr = MatrixLayout::RowMajor;
-    let problem = MatmulProblem::from_parameters(
-        m,
-        n,
-        k,
-        shape![b],
-        shape![b],
-        tl,
-        tr,
-        MatrixLayout::RowMajor,
-        None,
-        None,
-        MatmulElems::new_deprecated::<MP>().as_global_elems(),
-        AddressType::U32,
-    );
-
-    for t in [(16, 16, 16)] {
-        for p in [(1, 1, 1)] {
-            for s in [(1, 1, 1)] {
-                let plane_dim = client.properties().hardware.plane_size_min;
-                let tiling = TilingScheme::builder()
-                    .with_tile_size(t.into())
-                    .with_partition_size(p.into())
-                    .with_stage_size(StageSize {
-                        m: s.0,
-                        n: s.1,
-                        k: s.2,
-                    })
-                    .build()
-                    .unwrap();
-                let hypercube = HypercubeBlueprint::builder()
-                    .global_order(GlobalOrder::default())
-                    .cube_count_strategy(CubeCountStrategy::Flattened)
-                    .build();
-                let blueprint =
-                    TilingBlueprint::builder(TileMatmulKind::Cmma, tiling, plane_dim, &problem)
-                        .partition_buffering(PartitionBuffering::Single)
-                        .hypercube_blueprint(hypercube)
-                        .loading_precompute_strategy(LoadingPrecomputeStrategy::Always)
-                        .build();
-                let result = run_one::<R, MP>(
-                    Default::default(),
-                    Strategy::SimpleCyclicCmma(BlueprintStrategy::Forced(blueprint.clone())),
-                    &problem,
-                );
-                if let Ok((duration, tflops)) = result {
-                    let key = tflops * 1000000.0;
-                    algos.insert(key as u64, (duration, blueprint, tflops));
-                }
-            }
-        }
-    }
-
-    for (_, (duration, selection, tflops)) in algos.iter() {
-        println!("==== TFLOPS: {tflops:?} ====");
-        println!("Selection: {selection:?}");
-        println!("Times: {duration}");
-        println!("====================");
-    }
-}
-
-#[allow(unused)]
-fn run_algos_unit<R: Runtime, MP: MatmulPrecision>() {
-    let client = R::client(&Default::default());
-
-    println!("Simple Unit Min");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::SimpleUnit(BlueprintStrategy::Inferred(SimpleUnitSelectionArgs {
-            tile_size: TileSizeSelection::MinTileSize,
-        })),
-    );
-
-    println!("Simple Unit Max");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::SimpleUnit(BlueprintStrategy::Inferred(SimpleUnitSelectionArgs {
-            tile_size: TileSizeSelection::MaxTileSize,
-        })),
-    );
-
-    println!("Double Unit Min");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::DoubleUnit(BlueprintStrategy::Inferred(DoubleUnitSelectionArgs {
-            tile_size: TileSizeSelection::MinTileSize,
-        })),
-    );
-
-    println!("Double Unit Max");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::DoubleUnit(BlueprintStrategy::Inferred(DoubleUnitSelectionArgs {
-            tile_size: TileSizeSelection::MaxTileSize,
-        })),
-    );
-}
-
-#[allow(unused)]
-fn run_algos_wmma<R: Runtime, MP: MatmulPrecision>() {
-    let client = R::client(&Default::default());
-
-    println!("Simple");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::SimpleCyclicCmma(BlueprintStrategy::Inferred(SimpleArgs {
-            multi_rows: false,
-            ..Default::default()
-        })),
-    );
-
-    println!("Simple multi rows");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::SimpleCyclicCmma(BlueprintStrategy::Inferred(SimpleArgs {
-            multi_rows: true,
-            ..Default::default()
-        })),
-    );
-
-    println!("Double Buffering");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::DoubleTilewiseCmma(BlueprintStrategy::Inferred(DoubleBufferingArgs {
-            specialized: false,
-            ..Default::default()
-        })),
-    );
-
-    println!("Double Buffering Specialized");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::DoubleTilewiseCmma(BlueprintStrategy::Inferred(DoubleBufferingArgs {
-            specialized: true,
-            ..Default::default()
-        })),
-    );
-
-    println!("Double Buffering Ordered");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::OrderedDoubleCmma(BlueprintStrategy::Inferred(OrderedSelectionArgs {
-            row_count: Some(8),
-            rows_per_plane: Some(2),
-            partition_k: Some(2),
-            ..Default::default()
-        })),
-    );
-}
-
-#[allow(unused)]
-fn run_algos_mma<R: Runtime, MP: MatmulPrecision>() {
-    let client = R::client(&Default::default());
-
-    // println!("Specialized TMA");
-    // run::<R, MP>(
-    //     Default::default(),
-    //     matmul::Strategy::Specialized {
-    //         read_strategy: AsyncPartialReadingStrategy::Tma,
-    //         selection: Selection::Inferred(()),
-    //         tile_kind: AcceleratedTileKind::Mma,
-    //     },
-    // );
-
-    // println!("Specialized Cyclic");
-    // run::<R, MP>(
-    //     Default::default(),
-    //     matmul::Strategy::Specialized {
-    //         read_strategy: AsyncPartialReadingStrategy::Cyclic,
-    //         selection: Selection::Inferred(()),
-    //         tile_kind: AcceleratedTileKind::Mma,
-    //     },
-    // );
-
-    println!("Specialized Strided");
-    run::<R, MP>(
-        Default::default(),
-        Strategy::SpecializedStridedMma(BlueprintStrategy::Inferred(().into())),
-    );
-}
-
-#[allow(unused)]
-fn run_benches<R: Runtime, MP: MatmulPrecision>() {
-    // run_grid_search::<R, MP>();
-    // run_algos_unit::<R, MP>();
-    run_algos_wmma::<R, MP>();
-    // run_algos_mma::<R, MP>();
-}
+use benchmarks::gemm;
+use cubecl::benchmark::{BenchmarkDurations, TimingMethod};
 
 fn main() {
-    // run_benches::<cubecl::TestRuntime, half::f16>();
-    run_benches::<cubecl::TestRuntime, f32>();
+    for problem in gemm::problems() {
+        for strategy in gemm::strategies() {
+            println!("---- {} / {} ----", strategy.label, problem.label);
+            match gemm::run(&strategy.id, &problem.id, 10) {
+                Ok(samples) => {
+                    let durations = BenchmarkDurations {
+                        timing_method: TimingMethod::System,
+                        durations: samples.durations,
+                    };
+                    println!("{durations}");
+                }
+                Err(err) => println!("error: {err}"),
+            }
+        }
+    }
 }
