+++ benchmarks/benches/memcpy_async.rs
@@ -1,881 +1,20 @@
-use cubecl::{
-    benchmark::{Benchmark, ProfileDuration, TimingMethod},
-    frontend::Float,
-    future,
-    prelude::{barrier::Barrier, *},
-    std::tensor::TensorHandle,
-};
-use cubek::random::random_uniform;
-use std::marker::PhantomData;
-
-#[cube]
-trait ComputeTask: Send + Sync + 'static {
-    fn compute<E: Float, N: Size>(
-        input: &Slice<Vector<E, N>>,
-        acc: &mut Array<Vector<E, N>>,
-        #[comptime] config: Config,
-    );
-
-    fn to_output<E: Float, N: Size>(
-        acc: &mut Array<Vector<E, N>>,
-        output: &mut SliceMut<Vector<E, N>>,
-        #[comptime] config: Config,
-    );
-}
-
-#[derive(CubeType)]
-struct DummyCompute {}
-#[cube]
-impl ComputeTask for DummyCompute {
-    fn compute<E: Float, N: Size>(
-        input: &Slice<Vector<E, N>>,
-        acc: &mut Array<Vector<E, N>>,
-        #[comptime] config: Config,
-    ) {
-        // An offset to make sure units need the data loaded by other units
-        let offset = 256;
-
-        let position = (UNIT_POS as usize * config.acc_len + offset) % config.smem_size;
-        for i in 0..config.acc_len {
-            acc[i] += input[position + i];
-        }
-    }
-
-    fn to_output<E: Float, N: Size>(
-        acc: &mut Array<Vector<E, N>>,
-        output: &mut SliceMut<Vector<E, N>>,
-        #[comptime] config: Config,
-    ) {
-        let position = UNIT_POS as usize * config.acc_len;
-        for i in 0..config.acc_len {
-            acc[i] += output[position + i];
-        }
-    }
-}
-
-#[cube]
-trait CopyStrategy: Send + Sync + 'static {
-    type Barrier: CubeType + Copy + Clone;
-
-    fn barrier() -> Self::Barrier;
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    );
-
-    fn wait(_barrier: Self::Barrier);
-}
-
-#[derive(CubeType)]
-/// All units perform all loads, duplicatively
-/// Not async
-struct DummyCopy {}
-#[cube]
-impl CopyStrategy for DummyCopy {
-    type Barrier = ();
-
-    fn barrier() -> Self::Barrier {}
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        _barrier: Self::Barrier,
-        #[comptime] _config: Config,
-    ) {
-        for i in 0..source.len() {
-            destination[i] = source[i];
-        }
-    }
-
-    fn wait(_barrier: Self::Barrier) {
-        sync_cube();
-    }
-}
-
-#[derive(CubeType)]
-/// Units worked in a coalesced manner to share the loading
-/// Not async
-struct CoalescedCopy {}
-#[cube]
-impl CopyStrategy for CoalescedCopy {
-    type Barrier = ();
-
-    fn barrier() -> Self::Barrier {}
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        _barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let num_units = config.num_planes * config.plane_dim;
-        let num_copies_per_unit = source.len() as u32 / num_units;
-        for i in 0..num_copies_per_unit {
-            let pos = UNIT_POS + i * num_units;
-            destination[pos as usize] = source[pos as usize];
-        }
-    }
-
-    fn wait(_barrier: Self::Barrier) {
-        sync_cube();
-    }
-}
-
-#[derive(CubeType)]
-/// All units will call one slice that encompasses all data
-/// using a per-unit memcpy_async
-///
-/// Probably very slow as all units perform the copy duplicatively
-struct MemcpyAsyncSingleSliceDuplicatedAll {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSingleSliceDuplicatedAll {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] _config: Config,
-    ) {
-        barrier.memcpy_async(source, destination)
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// One unit is elected to call one slice that encompasses all data
-/// using a per-unit memcpy_async
-///
-/// Probably very slow as one unit does everything
-struct MemcpyAsyncSingleSliceElected {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSingleSliceElected {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] _config: Config,
-    ) {
-        if UNIT_POS == 0 {
-            barrier.memcpy_async(source, destination)
-        }
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// One unit is elected to call one slice that encompasses all data
-/// using a cooperative memcpy_async
-///
-/// Probably the best for contiguous data
-struct MemcpyAsyncSingleSliceElectedCooperative {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSingleSliceElectedCooperative {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] _config: Config,
-    ) {
-        if UNIT_POS == 0 {
-            barrier.memcpy_async(source, destination)
-        }
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// The slice is divided by the number of planes, then each subslice
-/// is handled by a plane.
-/// In this case all units of the plane call memcpy
-struct MemcpyAsyncSplitPlaneDuplicatedUnit {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSplitPlaneDuplicatedUnit {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let sub_length = source.len() as u32 / config.num_planes;
-        let start = UNIT_POS_Y * sub_length;
-        let end = start + sub_length;
-
-        barrier.memcpy_async(
-            &source.slice(start as usize, end as usize),
-            &mut destination.slice_mut(start as usize, end as usize),
-        )
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// The slice is divided by the number of planes, then each subslice
-/// is handled by a plane.
-/// In this case one unit of the plane is elected to call memcpy
-struct MemcpyAsyncSplitPlaneElectedUnit {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSplitPlaneElectedUnit {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let sub_length = source.len() as u32 / config.num_planes;
-        let start = UNIT_POS_Y * sub_length;
-        let end = start + sub_length;
-
-        if UNIT_POS_X == 0 {
-            barrier.memcpy_async(
-                &source.slice(start as usize, end as usize),
-                &mut destination.slice_mut(start as usize, end as usize),
-            )
-        }
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// The slice is divided by the number of planes (arbitrartily), but each
-/// unit calls each slice
-///
-/// Probably the worst
-struct MemcpyAsyncSplitDuplicatedAll {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSplitDuplicatedAll {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let sub_length = source.len() as u32 / config.num_planes;
-        for i in 0..config.num_planes {
-            let start = i * sub_length;
-            let end = start + sub_length;
-
-            barrier.memcpy_async(
-                &source.slice(start as usize, end as usize),
-                &mut destination.slice_mut(start as usize, end as usize),
-            )
-        }
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// The slice is divided by the number of planes n (arbitrarily)
-/// Then the first n units in absolute each perform a call
-struct MemcpyAsyncSplitLargeUnitWithIdle {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSplitLargeUnitWithIdle {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let sub_length = source.len() as u32 / config.num_planes;
-
-        if UNIT_POS < config.num_planes {
-            let start = UNIT_POS * sub_length;
-            let end = start + sub_length;
-
-            barrier.memcpy_async(
-                &source.slice(start as usize, end as usize),
-                &mut destination.slice_mut(start as usize, end as usize),
-            )
-        }
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// The slice is divided into single elements.
-/// then all units make coalesced calls for a single element,
-/// looping cyclicly until the end
-struct MemcpyAsyncSplitSmallUnitCoalescedLoop {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSplitSmallUnitCoalescedLoop {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let num_units = config.num_planes * config.plane_dim;
-        let num_loops = source.len() as u32 / num_units;
-
-        for i in 0..num_loops {
-            let start = UNIT_POS + i * num_units;
-            let end = start + 1;
-
-            barrier.memcpy_async(
-                &source.slice(start as usize, end as usize),
-                &mut destination.slice_mut(start as usize, end as usize),
-            )
-        }
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(CubeType)]
-/// The slice is divided by the number of units,
-/// then all units make one coalesced call of the length of the subslices
-///
-/// Probably the fastest for a non contiguous series of contiguous slices
-struct MemcpyAsyncSplitMediumUnitCoalescedOnce {}
-#[cube]
-impl CopyStrategy for MemcpyAsyncSplitMediumUnitCoalescedOnce {
-    type Barrier = Shared<Barrier>;
-
-    fn barrier() -> Self::Barrier {
-        Barrier::shared(CUBE_DIM, UNIT_POS == 0u32)
-    }
-
-    fn memcpy<E: Float, N: Size>(
-        source: &Slice<Vector<E, N>>,
-        destination: &mut SliceMut<Vector<E, N>>,
-        barrier: Self::Barrier,
-        #[comptime] config: Config,
-    ) {
-        let sub_length = source.len() as u32 / (config.num_planes * config.plane_dim);
-
-        let start = UNIT_POS * sub_length;
-        let end = start + sub_length;
-
-        barrier.memcpy_async(
-            &source.slice(start as usize, end as usize),
-            &mut destination.slice_mut(start as usize, end as usize),
-        )
-    }
-
-    fn wait(barrier: Self::Barrier) {
-        barrier.arrive_and_wait();
-    }
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
-struct Config {
-    plane_dim: u32,
-    num_planes: u32,
-    smem_size: usize,
-    acc_len: usize,
-    double_buffering: bool,
-}
-
-#[cube(launch_unchecked)]
-fn memcpy_test<E: Float, N: Size, Cpy: CopyStrategy, Cpt: ComputeTask>(
-    input: &Tensor<Vector<E, N>>,
-    output: &mut Tensor<Vector<E, N>>,
-    #[comptime] config: Config,
-) {
-    if config.double_buffering {
-        memcpy_test_single_buffer::<E, N, Cpy, Cpt>(input, output, config);
-    } else {
-        memcpy_test_double_buffer::<E, N, Cpy, Cpt>(input, output, config);
-    }
-}
-
-#[cube]
-fn memcpy_test_single_buffer<E: Float, N: Size, Cpy: CopyStrategy, Cpt: ComputeTask>(
-    input: &Tensor<Vector<E, N>>,
-    output: &mut Tensor<Vector<E, N>>,
-    #[comptime] config: Config,
-) {
-    let data_count = input.shape(0);
-    let mut acc = Array::<Vector<E, N>>::new(config.acc_len);
-    let num_iterations = data_count.div_ceil(config.smem_size);
-
-    let mut smem = SharedMemory::<Vector<E, N>>::new(config.smem_size);
-    let barrier = Cpy::barrier();
-
-    for i in 0..num_iterations {
-        let start = i * config.smem_size;
-        let end = start + config.smem_size;
-
-        Cpy::memcpy(
-            &input.slice(start, end),
-            &mut smem.to_slice_mut(),
-            barrier,
-            config,
-        );
-
-        Cpy::wait(barrier);
-
-        Cpt::compute(&smem.to_slice(), &mut acc, config);
-    }
-
-    Cpy::wait(barrier);
-    Cpt::compute(&smem.to_slice(), &mut acc, config);
-    Cpt::to_output(&mut acc, &mut output.to_slice_mut(), config);
-}
-
-#[cube]
-fn memcpy_test_double_buffer<E: Float, N: Size, Cpy: CopyStrategy, Cpt: ComputeTask>(
-    input: &Tensor<Vector<E, N>>,
-    output: &mut Tensor<Vector<E, N>>,
-    #[comptime] config: Config,
-) {
-    let data_count = input.shape(0);
-    let mut smem1 = SharedMemory::<Vector<E, N>>::new(config.smem_size);
-    let mut smem2 = SharedMemory::<Vector<E, N>>::new(config.smem_size);
-    let mut acc = Array::<Vector<E, N>>::new(config.acc_len);
-    let num_iterations = data_count.div_ceil(config.smem_size);
-
-    let barrier1 = Cpy::barrier();
-    let barrier2 = Cpy::barrier();
-
-    for i in 0..num_iterations {
-        let start = i * config.smem_size;
-        let end = if start + config.smem_size < data_count {
-            start + config.smem_size
-        } else {
-            data_count
-        };
-
-        if i % 2 == 0 {
-            Cpy::memcpy(
-                &input.slice(start, end),
-                &mut smem1.to_slice_mut(),
-                barrier1,
-                config,
-            );
-            if i > 0 {
-                Cpy::wait(barrier2);
-                Cpt::compute(&smem2.to_slice(), &mut acc, config);
-            }
-        } else {
-            Cpy::memcpy(
-                &input.slice(start, end),
-                &mut smem2.to_slice_mut(),
-                barrier2,
-                config,
-            );
-
-            Cpy::wait(barrier1);
-            Cpt::compute(&smem1.to_slice(), &mut acc, config);
-        }
-    }
-
-    Cpy::wait(barrier2);
-    Cpt::compute(&smem2.to_slice(), &mut acc, config);
-
-    Cpt::to_output(&mut acc, &mut output.to_slice_mut(), config);
-}
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
-#[allow(unused)]
-/// Strategy in use. Refer to the struct of the same name as the variant for doc.
-enum CopyStrategyEnum {
-    DummyCopy,
-    CoalescedCopy,
-    MemcpyAsyncSingleSliceDuplicatedAll,
-    MemcpyAsyncSingleSliceElected,
-    MemcpyAsyncSingleSliceElectedCooperative,
-    MemcpyAsyncSplitPlaneDuplicatedUnit,
-    MemcpyAsyncSplitPlaneElectedUnit,
-    MemcpyAsyncSplitDuplicatedAll,
-    MemcpyAsyncSplitLargeUnitWithIdle,
-    MemcpyAsyncSplitSmallUnitCoalescedLoop,
-    MemcpyAsyncSplitMediumUnitCoalescedOnce,
-}
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash)]
-#[allow(unused)]
-enum ComputeTaskEnum {
-    Dummy,
-}
-
-fn launch_ref<R: Runtime, E: Float>(
-    strategy: CopyStrategyEnum,
-    client: &ComputeClient<R>,
-    input: TensorBinding<R>,
-    output: TensorBinding<R>,
-    smem_size: usize,
-    double_buffering: bool,
-) {
-    let cube_count = CubeCount::Static(1, 1, 1);
-    let plane_dim = 32;
-    let num_planes = 8;
-    let cube_dim = CubeDim::new_2d(plane_dim, num_planes);
-    let config = Config {
-        plane_dim,
-        num_planes,
-        smem_size,
-        acc_len: smem_size / (plane_dim * num_planes) as usize,
-        double_buffering,
-    };
-
-    unsafe {
-        match strategy {
-            CopyStrategyEnum::DummyCopy => {
-                memcpy_test::launch_unchecked::<E, DummyCopy, DummyCompute, R>(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::CoalescedCopy => {
-                memcpy_test::launch_unchecked::<E, CoalescedCopy, DummyCompute, R>(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSingleSliceDuplicatedAll => {
-                memcpy_test::launch_unchecked::<
-                    E,
-                    MemcpyAsyncSingleSliceDuplicatedAll,
-                    DummyCompute,
-                    R,
-                >(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSingleSliceElected => {
-                memcpy_test::launch_unchecked::<E, MemcpyAsyncSingleSliceElected, DummyCompute, R>(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSingleSliceElectedCooperative => {
-                memcpy_test::launch_unchecked::<
-                    E,
-                    MemcpyAsyncSingleSliceElectedCooperative,
-                    DummyCompute,
-                    R,
-                >(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSplitPlaneDuplicatedUnit => {
-                memcpy_test::launch_unchecked::<
-                    E,
-                    MemcpyAsyncSplitPlaneDuplicatedUnit,
-                    DummyCompute,
-                    R,
-                >(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSplitPlaneElectedUnit => {
-                memcpy_test::launch_unchecked::<E, MemcpyAsyncSplitPlaneElectedUnit, DummyCompute, R>(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSplitDuplicatedAll => {
-                memcpy_test::launch_unchecked::<E, MemcpyAsyncSplitDuplicatedAll, DummyCompute, R>(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSplitLargeUnitWithIdle => {
-                memcpy_test::launch_unchecked::<E, MemcpyAsyncSplitLargeUnitWithIdle, DummyCompute, R>(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSplitSmallUnitCoalescedLoop => {
-                memcpy_test::launch_unchecked::<
-                    E,
-                    MemcpyAsyncSplitSmallUnitCoalescedLoop,
-                    DummyCompute,
-                    R,
-                >(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-            CopyStrategyEnum::MemcpyAsyncSplitMediumUnitCoalescedOnce => {
-                memcpy_test::launch_unchecked::<
-                    E,
-                    MemcpyAsyncSplitMediumUnitCoalescedOnce,
-                    DummyCompute,
-                    R,
-                >(
-                    client,
-                    cube_count,
-                    cube_dim,
-                    1,
-                    input.into_tensor_arg(),
-                    output.into_tensor_arg(),
-                    config,
-                )
-            }
-        };
-    }
-}
-
-impl<R: Runtime, E: Float> Benchmark for MemcpyAsyncBench<R, E> {
-    type Input = (TensorHandle<R>, TensorHandle<R>);
-    type Output = ();
-
-    fn prepare(&self) -> Self::Input {
-        let client = R::client(&self.device);
-
-        let a = TensorHandle::empty(
-            &client,
-            vec![self.data_count],
-            E::as_type_native_unchecked(),
-        );
-        random_uniform(
-            &client,
-            0.,
-            1.,
-            a.clone().binding(),
-            E::as_type_native_unchecked().storage_type(),
-        )
-        .unwrap();
-        let b = TensorHandle::empty(
-            &client,
-            vec![self.window_size],
-            E::as_type_native_unchecked(),
-        );
-        random_uniform(
-            &client,
-            0.,
-            1.,
-            b.clone().binding(),
-            E::as_type_native_unchecked().storage_type(),
-        )
-        .unwrap();
-
-        (a, b)
-    }
-
-    fn execute(&self, args: Self::Input) -> Result<(), String> {
-        let smem_size = args.1.shape()[0];
-        launch_ref::<R, E>(
-            self.strategy,
-            &self.client,
-            args.0.binding(),
-            args.1.binding(),
-            smem_size,
-            self.double_buffering,
-        );
-        Ok(())
-    }
-
-    fn name(&self) -> String {
-        let client = R::client(&self.device);
-
-        format!(
-            "memcpy_async-{}-{}-{:?}",
-            R::name(&client),
-            E::as_type_native_unchecked(),
-            self.strategy
-        )
-        .to_lowercase()
-    }
-
-    fn sync(&self) {
-        future::block_on(self.client.sync()).unwrap()
-    }
-
-    fn profile(&self, args: Self::Input) -> Result<ProfileDuration, String> {
-        self.client
-            .profile(|| self.execute(args), "memcpy-async-bench")
-            .map(|it| it.1)
-            .map_err(|it| format!("{it:?}"))
-    }
-}
-
-#[allow(dead_code)]
-struct MemcpyAsyncBench<R: Runtime, E> {
-    data_count: usize,
-    window_size: usize,
-    strategy: CopyStrategyEnum,
-    double_buffering: bool,
-    device: R::Device,
-    client: ComputeClient<R>,
-    _e: PhantomData<E>,
-}
-
-#[allow(dead_code)]
-fn run<R: Runtime, E: Float>(device: R::Device, strategy: CopyStrategyEnum) {
-    let client = R::client(&device);
-
-    let data_count = 10000000;
-    let window_size = 2048;
-
-    let bench = MemcpyAsyncBench::<R, E> {
-        data_count,
-        window_size,
-        strategy,
-        double_buffering: true,
-        client: client.clone(),
-        device: device.clone(),
-        _e: PhantomData,
-    };
-    println!("Data count: {data_count:?}, strategy: {strategy:?}");
-    println!("{}", bench.name());
-    println!("{}", bench.run(TimingMethod::Device).unwrap());
-}
+use benchmarks::memcpy_async;
+use cubecl::benchmark::{BenchmarkDurations, TimingMethod};
 
 fn main() {
-    run::<cubecl::TestRuntime, f32>(Default::default(), CopyStrategyEnum::DummyCopy);
-    run::<cubecl::TestRuntime, f32>(Default::default(), CopyStrategyEnum::CoalescedCopy);
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSingleSliceDuplicatedAll,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSingleSliceElected,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSingleSliceElectedCooperative,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSplitPlaneDuplicatedUnit,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSplitPlaneElectedUnit,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSplitDuplicatedAll,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSplitLargeUnitWithIdle,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSplitSmallUnitCoalescedLoop,
-    );
-    run::<cubecl::TestRuntime, f32>(
-        Default::default(),
-        CopyStrategyEnum::MemcpyAsyncSplitMediumUnitCoalescedOnce,
-    );
+    for problem in memcpy_async::problems() {
+        for strategy in memcpy_async::strategies() {
+            println!("---- {} / {} ----", strategy.label, problem.label);
+            match memcpy_async::run(&strategy.id, &problem.id, 10) {
+                Ok(samples) => {
+                    let durations = BenchmarkDurations {
+                        timing_method: TimingMethod::Device,
+                        durations: samples.durations,
+                    };
+                    println!("{durations}");
+                }
+                Err(err) => println!("error: {err}"),
+            }
+        }
+    }
 }
