+++ benchmarks/benches/unary.rs
@@ -1,111 +1,20 @@
-use cubecl::{
-    benchmark::{Benchmark, ProfileDuration, TimingMethod},
-    calculate_cube_count_elemwise, frontend, future,
-    prelude::*,
-    std::tensor::TensorHandle,
-};
-use cubek::random::random_uniform;
-use std::marker::PhantomData;
+use benchmarks::unary;
+use cubecl::benchmark::{BenchmarkDurations, TimingMethod};
 
-#[cube(launch)]
-fn execute<F: Float>(lhs: &Tensor<F>, rhs: &Tensor<F>, out: &mut Tensor<F>) {
-    if ABSOLUTE_POS < out.len() {
-        for i in 0..256u32 {
-            if i % 2 == 0 {
-                out[ABSOLUTE_POS] -= F::cos(lhs[ABSOLUTE_POS] * rhs[ABSOLUTE_POS]);
-            } else {
-                out[ABSOLUTE_POS] += F::cos(lhs[ABSOLUTE_POS] * rhs[ABSOLUTE_POS]);
+fn main() {
+    for problem in unary::problems() {
+        for strategy in unary::strategies() {
+            println!("---- {} / {} ----", strategy.label, problem.label);
+            match unary::run(&strategy.id, &problem.id, 10) {
+                Ok(samples) => {
+                    let durations = BenchmarkDurations {
+                        timing_method: TimingMethod::Device,
+                        durations: samples.durations,
+                    };
+                    println!("{durations}");
+                }
+                Err(err) => println!("error: {err}"),
             }
         }
     }
-}
-
-impl<R: Runtime, E: Float> Benchmark for UnaryBench<R, E> {
-    type Input = (TensorHandle<R>, TensorHandle<R>, TensorHandle<R>);
-    type Output = ();
-
-    fn prepare(&self) -> Self::Input {
-        let client = R::client(&self.device);
-        let elem = E::as_type_native_unchecked();
-
-        let lhs = TensorHandle::empty(&client, self.shape.clone(), elem);
-        random_uniform(&client, 0., 1., lhs.clone().binding(), elem.storage_type()).unwrap();
-        let rhs = TensorHandle::empty(&client, self.shape.clone(), elem);
-        random_uniform(&client, 0., 1., rhs.clone().binding(), elem.storage_type()).unwrap();
-        let out = TensorHandle::empty(&client, self.shape.clone(), elem);
-        random_uniform(&client, 0., 1., out.clone().binding(), elem.storage_type()).unwrap();
-
-        (lhs, rhs, out)
-    }
-
-    fn execute(&self, (lhs, rhs, out): Self::Input) -> Result<(), String> {
-        let num_elems = out.shape().num_elements();
-
-        let working_units = num_elems / self.vectorization;
-        let cube_dim = CubeDim::new(&self.client, working_units);
-        let cube_count = calculate_cube_count_elemwise(&self.client, working_units, cube_dim);
-
-        execute::launch::<E, R>(
-            &self.client,
-            cube_count,
-            cube_dim,
-            lhs.into_arg(),
-            rhs.into_arg(),
-            out.into_arg(),
-        );
-
-        Ok(())
-    }
-
-    fn name(&self) -> String {
-        let client = R::client(&self.device);
-
-        format!(
-            "unary-{}-{}-{:?}",
-            R::name(&client),
-            E::as_type_native_unchecked(),
-            self.vectorization
-        )
-        .to_lowercase()
-    }
-
-    fn sync(&self) {
-        future::block_on(self.client.sync()).unwrap()
-    }
-
-    fn profile(&self, args: Self::Input) -> Result<ProfileDuration, String> {
-        self.client
-            .clone()
-            .profile(|| self.execute(args), "unary-bench")
-            .map(|it| it.1)
-            .map_err(|it| format!("{it:?}"))
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Clone)]
-struct UnaryBench<R: Runtime, E> {
-    shape: Vec<usize>,
-    vectorization: VectorSize,
-    device: R::Device,
-    client: ComputeClient<R>,
-    _e: PhantomData<E>,
-}
-
-#[allow(dead_code)]
-fn run<R: Runtime, E: frontend::Float>(device: R::Device, vectorization: VectorSize) {
-    let client = R::client(&device);
-    let bench = UnaryBench::<R, E> {
-        shape: vec![32, 512, 2048],
-        vectorization,
-        client,
-        device,
-        _e: PhantomData,
-    };
-    println!("{}", bench.name());
-    println!("{}", bench.run(TimingMethod::Device).unwrap());
-}
-
-fn main() {
-    run::<cubecl::TestRuntime, f32>(Default::default(), 4);
 }
