+++ benchmarks/src/attention/benchmark.rs
@@ -67,7 +67,7 @@ pub fn run_on<R: Runtime, AP: AttentionPrecision>(
         .map_err(|e| format!("benchmark failed: {e}"))?
         .durations;
 
-    Ok(RunSamples { durations })
+    Ok(RunSamples::new(durations))
 }
 
 struct AttentionBench<R: Runtime, AP> {
