+++ benchmarks/src/conv2d/mod.rs
@@ -12,10 +12,10 @@ pub struct Category;
 
 impl BenchmarkCategory for Category {
     fn id(&self) -> &'static str {
-        "attention"
+        "conv2d"
     }
     fn label(&self) -> &'static str {
-        "Attention"
+        "Conv2d"
     }
     fn strategies(&self) -> Vec<ItemDescriptor> {
         strategies()
