+++ benchmarks/src/fft/mod.rs
@@ -12,10 +12,10 @@ pub struct Category;
 
 impl BenchmarkCategory for Category {
     fn id(&self) -> &'static str {
-        "attention"
+        "fft"
     }
     fn label(&self) -> &'static str {
-        "Attention"
+        "FFT"
     }
     fn strategies(&self) -> Vec<ItemDescriptor> {
         strategies()
