+++ benchmarks/src/gemm/mod.rs
@@ -12,10 +12,10 @@ pub struct Category;
 
 impl BenchmarkCategory for Category {
     fn id(&self) -> &'static str {
-        "attention"
+        "gemm"
     }
     fn label(&self) -> &'static str {
-        "Attention"
+        "GEMM"
     }
     fn strategies(&self) -> Vec<ItemDescriptor> {
         strategies()
