+++ benchmarks/src/gemv/mod.rs
@@ -12,10 +12,10 @@ pub struct Category;
 
 impl BenchmarkCategory for Category {
     fn id(&self) -> &'static str {
-        "attention"
+        "gemv"
     }
     fn label(&self) -> &'static str {
-        "Attention"
+        "GEMV"
     }
     fn strategies(&self) -> Vec<ItemDescriptor> {
         strategies()
