+++ benchmarks/src/unary/mod.rs
@@ -12,10 +12,10 @@ pub struct Category;
 
 impl BenchmarkCategory for Category {
     fn id(&self) -> &'static str {
-        "attention"
+        "unary"
     }
     fn label(&self) -> &'static str {
-        "Attention"
+        "Unary"
     }
     fn strategies(&self) -> Vec<ItemDescriptor> {
         strategies()
